refactor: modularize simulator constants and helpers
